alice/beggins/internal/nirvana/beggins/requirements.dev.txt
attrs==20.2.0
h5py==2.10.0
numpy==1.18.5
tensorflow==1.15.4
tensorboard==1.15.0
protobuf==3.13.0
six==1.15.0
PyYAML==6.0
click==8.0.3
